Internals

Circuit Elements internals

Qurt.Elements.Q1GateBlocks — Constant

Q1GateBlocks

A Tuple of block numbers of Element that contain one-qubit gates.

Qurt.Elements.Element — Type

Element <: BlockEnum

An enum representing circuit elements. These are gates, instructions, I/O nodes.

The values of Element are in one-to-one correspondence with inteters and are arranged in blocks that are assigned semantics corresponding to classes of circuit elements. For example this first block is for one-qubit, parameterless, builtin gates.

Qurt.Elements.OpBlock — Type

OpBlock <: BlockEnum

An enum that gives names to the blocks of values of Element.

Qurt.Elements.ParamElement — Type

ParamElement{ParamsT}

Struct representing a parameterized circuit element.

ParamElement has two fields; the element and a tuple of parameters. This is not the fundamental structure representing parameterized circuit elements. Rather, it is a convenient grouping when both of these properties are needed but not others, namely the wires.

Qurt.Elements.WiresElement — Type

WiresElement{QuWiresT,ClWiresT}

Structure holding a circuit element element::Element and the quantum wires and classical wires in a circuit that it is connected to.

Qurt.Elements.WiresParamElement — Type

WiresParamElement{QuWiresT,ClWiresT,ParamsT}

Structure holding a circuit element element::Element, its parameter, and the quantum wires and classical wires in a circuit that it is connected to.

Qurt.Angle.isapprox_turn — Method

isapprox_turn(x::ParamElement, y::ParamElement; kw...)

Return true if x and y are approximately equal. The element types must be equal. Angle.isapprox_turn must return true element-wise on the parameters.

Qurt.Elements._check_begin_block — Method

_check_begin_block(syms)

Get the symbols from an expression captured by our macros.

syms is either a sequence of symbols, or a sequence of symbols wrapped in a begin-end block. In either case, return just the sequence of symbols.

Qurt.Elements._get_qsyms — Method

_get_qsyms(syms)

Get the symbols from the expression syms may contain extraneous items.

Removes a possible enclosing block, and filters out any LineNumberNode.

Qurt.Elements.inblocks — Method

inblocks(elem::Element, blocks)

Return true if elem is in any block in blocks.

Qurt.Elements.@new_elements — Macro

@new_elements BlockName sym1 sym2 ...

Add new circuit element symbols to the block of elements named BlockName.

Examples

julia> @new_elements MiscGates MyGate1 MyGate2

Circuits internals

Base.empty — Method

empty(qc::Circuit)

Return an object that is a copy of qc except that all circuit elements other than input and output nodes are not present.

Qurt.Circuits.__add_io_node_data! — Method

__add_io_node_data!(graph, nodes, nqubits, nclbits)

Add input and output nodes to nodes. Wires numbered 1 through nqubits are quantum wires. Wires numbered nqubits + 1 through nqubits + nclbits are classical wires.

Qurt.Circuits.check_param_table — Method

check_param_table(qc::Circuit)

Check that the (node, paramter position) pairs recorded in the table actually contain the recorded parameter. This check could fail if reindexing is not done properly.

Qurt.Interface.check — Method

check(qc::Circuit)

Throw an Exception if any of a few checks on the integrity of qc fail.

Qurt.Interface.count_op_elements — Method

count_op_elements(qc::Circuit)

Return the number of circuit elements that are not IO nodes. This should be the number that are operation or instruction nodes.

Examples

julia> import Qurt

julia> using Qurt.Circuits: Circuit, count_op_elements

julia> using Qurt.Builders: @build

julia> using Qurt.Elements: H, CX

julia> qc = Circuit(2);

julia> count_op_elements(qc)
0

julia> @build qc H(1) CX(1, 2);

julia> count_op_elements(qc)
2

Qurt.Interface.depth — Method

depth(qc::Circuit)

Compute the depth of qc.

Qurt.Interface.num_clbits — Method

num_clbits(qc::Circuit)

Return the number of classical bits in qc.

Qurt.Interface.num_parameters — Method

num_parameters(qc::Circuit)

Return the number of unique symbolic parameter expressions in use in gates in qc.

For example, if exactly t1, t2, and t1 - t2 are present, then num_parameters will return 3. Non-symbolic parameters, such as 1.5, do not count.

Qurt.Interface.num_qubits — Method

num_qubits(qc::Circuit)

Return the number of qubits in qc.

Qurt.NodeStructs.wireind — Method

wireind(circuit, node_ind, wire::Integer)

Return the index of wire number wire in the list of wires for node node_ind.

Wires internals

Qurt.WiresMod.Wires — Type

struct Wires

This struct tracks data needed to identify wires and their input and output vertices. When possible, we don't distinguish between classical and quantum wires in the DAG. So, we don't number them separately. Qiskit supports adding and removing classical wires after the circuit is constructed. It would be simpler if we did not support this. But we do. So we keep an array holding the ordinal wire numbers of the quantum wires and another for the classical wires. We also keep four arrays, one to store each of input and output vertices for each of quantum and classical wires. For efficiency, we store all of the input and output vertices in two more arrays.

These ordinal (or vertex) numbers are currently Int (usually Int64), but we could switch to Int32. Considering just the $n$ qubits, we need storage for $4n$ integers.

It would be useful to simplify this structure and reduce storage requirements.

Qurt.WiresMod.input_vertex — Method

input_vertex(w::Wires, wireind::Integer)

Return the circuit vertex holding the input Element for wire wireind.

Qurt.WiresMod.output_vertex — Method

output_vertex(w::Wires, wireind::Integer)

Return the circuit vertex holding the output Element for wire wireind.

Interface internals

Qurt.Interface.iscustomgate — Function

iscustomgate(gate)

Return true if the gate is tagged CustomGate::Element.

Nodes internals

Graphs.inneighbors — Method

inneighbors(circuit, node_ind::Integer, wire::Integer)

Return the node index connected to node_ind by incoming wire number wire.

Graphs.inneighbors — Method

inneighbors(circuit, node_ind::Integer)

Return collection of incoming neighbor nodes in wire order.

Nodes may appear more than once if they are connected by multiple wires.

Graphs.outneighbors — Method

outneighbors(circuit, node_ind::Integer, wire::Integer)

Return the node index connected to node_ind by outgoing wire number wire.

Graphs.outneighbors — Method

outneighbors(circuit, node_ind::Integer)

Return collection of outgoing neighbor nodes in wire order.

Nodes may appear more than once if they are connected by multiple wires.

Qurt.Interface.count_ops_vertices — Method

count_ops_vertices(nodes::ANodeArrays, vertices)

Return a count map of elements on vertices.

Qurt.Interface.getparam — Method

getparam(nodes::ANodeArrays, ind::Integer, pos::Integer)

The the posth parameter at node index ind in nodes.

Qurt.NodeStructs.find_nodes — Method

find_nodes(testfunc::F, nodes::NodeVector, fieldname::Symbol) where {F}
find_nodes(testfunc::F, nodes::NodeVector, fieldnames::Tuple) where {F}
find_nodes(testfunc::F, nodes::NodeVector, ::Val{fieldnames}) where {F, fieldnames}

Return a view of nodes filtered by testfunc.

testfunc must take a single argument. It will be passed an structure with fields fieldnames. Only nodes for which testfunc returns true will be kept.

Calling find_nodes with fewer fields is more performant.

Examples

Find two qubit operations.

julia> find_nodes(x -> x.numquwires == 2, nodes, :numquwires)

Qurt.NodeStructs.one_qubit_ops — Method

one_qubit_ops(nodes::ANodeArrays)

Return a view of nodes containing all with two qubit wires.

Qurt.NodeStructs.rewire_across_node! — Method

rewire_across_node!(nodes::ANodeArrays, vind::Integer)

Wire incoming neighbors of vind to outgoing neighbors of vind, preserving the order of wires on ports.

Qurt.NodeStructs.unpackwires — Method

unpackwires(wires, nqu::Integer)

Return a tuple of two tuples, where the first contains the quantum wires and the second contains the classical wires.

wires contains all of the wires, with quantum wires first.

Builders internals

GraphUtils internals