Qurt.jl

Qurt.jl

Qurt

Circuit Elements

Circuits

Interface

Nodes

Builders

Passes

CompoundGates

Wires

Angle

IOQDags

Index

Internals

Circuit Elements internals

Circuits internals

Wires internals

Interface internals

Nodes internals

Builders internals

GraphUtils internals

Qurt

Qurt.Qurt — Module

module Qurt

The toplevel module of the package Qurt for building and manipulating quantum circuits. Documentation for Qurt is found in submodules.

There is apparently no way to include documentation from extension modules. There are two extension modules

PythonCallExt This extension will be loaded if you add PythonCall to your environment and load it. It defines methods for to_qiskit and draw for Qurt.Circuits.Circuit. Some documentation has been added to Interface for this.

GraphPlotExt This extension will be loaded if you add GraphPlot to your environment and load it. It contains not-well-developed functions for drawing the Qurt.Circuits.Circuit as a DAG.

Circuit Elements

Qurt.Elements — Module

module Elements

Quantum circuit elements. These are enums, that is encoded essentially as integers. There is a struct below ParamElement that composes an Element with a parameter, or container of parameters. This is meant to be any kind of parameter, not just an angle. But it is no meant to carry around metadata that is unrelated to paramaterizing a parametric gate.

Qurt.Elements.X — Constant

X::Element

The X gate circuit element.

Qurt.Elements.Y — Constant

Y::Element

The Y gate circuit element.

Circuits

Qurt.Circuits — Module

module Circuits

This module defines the struct Circuit to represent circuits and includes low-level functions for manipulating them. Functions in this module are often forwarded to, or use, functions in the module NodeStructs.

Qurt.Circuits.Circuit — Type

Circuit(nqubits::Integer, nclbits::Integer=0; global_phase=0.0)

Create a circuit with nqubits qubits, nclbits clbits.

Pairs of input and output nodes connected by an edges are created for each quantum and classical bit.

Examples

julia> using Qurt.Circuits: Circuit

julia> Circuit(2, 2)
circuit {nq=2, ncl=2, nv=8, ne=4} Graphs.SimpleGraphs.SimpleDiGraph{Int64} Qurt.NodeStructs.Node{Int64}

Qurt.Circuits.Circuit — Type

Circuit

Structure for representing a quantum circuit as a DAG with data attached to vertices and edges.

Fields

graph – The DAG as a Graphs.DiGraph

nodes – Operations and other nodes on vertices

nqubits – Number of qubits.

nclbits – Number of classical bits.

The DAG is a Graphs.DiGraph, which maintains edge lists for forward and backward edges. An "operation" is associated with each vertex in the graph. Each vertex is identified by a positive integer. Each wire is identified by a positive integer.

The edge lists for vertex i are given by the ith element of the Vector of edge lists stored in the DAG.

The operation on vertex i is given by the ith element of the field nodes.

There is no meaning in the order of neighboring vertices in the edge lists, in fact they are sorted.

The number of wires is equal to nqubits + nclbits.

Qurt.Circuits.Circuit — Method

Circuit(;global_phase=0.0)

Create a circuit with no qubits, no clbits, and global phase equal to zero.

Qurt.Circuits.add_node! — Function

add_node!(qcircuit::Circuit, op::Element, wires::NTuple{<:Any, IntT},
               clwires=()) where {IntT <: Integer}

add_node!(qcircuit::Circuit, (op, params)::Tuple{Element, <:Any},
                   wires::NTuple{<:Any, IntT}, clwires=()) where {IntT <: Integer}

Add op or (op, params) to the back of qcircuit with the specified classical and quantum wires.

The new node is inserted between the output nodes and their current predecessor nodes.

Qurt.Circuits.compose — Function

compose(qc_to::Circuit, qc_from::Circuit, quwires=1:num_wires(qc_from))

Append qc_from to a copy of qc_to.

Qurt.Circuits.compose! — Function

compose!(qc_to::Circuit, qc_from::Circuit, wireorder=1:num_wires(qc_from))

Append qc_from to qc_to.

qc_from must be narrower than, that is have fewer wires than, qc_to. wireorder maps wires in qc_from to those in qc_to.

Qurt.Circuits.count_ops_longest_path — Method

count_ops_longest_path(qc::Circuit)

Return a count map of the circuit elements on a longest path in qc.

Qurt.Circuits.global_phase — Method

global_phase(qc::Circuit)

Return the global phase of qc.

Qurt.Circuits.insert_node! — Function

insert_node!(qcircuit::Circuit, op::Element, out_vertices, wires::NTuple{<:Any, IntT},
               clwires=()) where {IntT <: Integer}

insert_node!(qcircuit::Circuit, (op, params)::Tuple{Element, <:Any},
                   out_vertices, wires::NTuple{<:Any, IntT}, clwires=()) where {IntT <: Integer}

Insert op or (op, params) to qcircuit before out_vertices on wires and clwires.

op is wired into the circuit at pairs in zip((wires..., clwires...), out_vertices)

Qurt.Circuits.longest_path — Method

longest_path(qc::Circuit)

Compute a longest path of vertices in qc.

Qurt.Circuits.nodes — Method

nodes(qc::Circuit)

Return the nodes in the circuit.

Qurt.Circuits.num_tensor_factors — Method

num_tensor_factors(qc::Circuit)

Return the number of tensor factors in an operator representation of qc.

The meaning of this in the presence of classical components is unclear.

Examples

julia> using Qurt.Circuits: Circuit, num_tensor_factors

julia> num_tensor_factors(Circuit(3, 2))
5

Qurt.Circuits.param_map — Method

param_map(qc::Circuit)

Return the parameter map for qc.

Qurt.Circuits.param_table — Method

param_table(qc::Circuit)

Return the parameter table for qc.

Qurt.Circuits.predecessors — Method

predecessors(qc::Circuit, vert)

Return the predecessors of vert in qc. This does not return a copy, so mutation will mutate the graph as well.

Qurt.Circuits.quantum_successors — Method

quantum_successors(qc::Circuit, vert)

Return the successors of vert in qc that are connnected by at least one quantum wire.

The return value is may or may not be a copy of node data.

Qurt.Circuits.remove_block! — Function

remove_block!(qc::Circuit, vinds, [vmap])

Remove the nodes in the block given by collection vinds and connect incoming and outgoing neighbors of the block on each wire. It is assumed that all incoming wires to the block connect to the first element of vinds and all the outgoing wires from the block connect to the last element of vinds. And the interior elements of vinds are not connected to any nodes outside of the block.

Qurt.Circuits.remove_blocks! — Method

remove_blocks!(qc::Circuit, blocks)

Remove vertices in blocks from qc, where each block in blocks is a Vector of vertices.

It is assumed that each block is as described in remove_block!.

Qurt.Circuits.remove_node! — Method

remove_node!(qc::Circuit, vind::Integer)

Remove the node at vertex index vind and connect incoming and outgoing neighbors on each wire.

Qurt.Circuits.successors — Method

successors(qc::Circuit, vert)

Return the successors of vert in qc. This does not return a copy, so mutation will mutate the graph as well.

Qurt.Circuits.topological_nodes — Method

topological_nodes(qc::Circuit)::AbstractVector{<:Node}

Return a topologically sorted vector of the nodes.

The returned data is a vector-of-structs view of the underlying data.

Qurt.Circuits.topological_vertices — Method

topological_vertices(qc::Circuit)::Vector{<:Integer}

Return a topologically sorted vector of the vertices.

Qurt.NodeStructs.wireelements — Function

wireelements(qc::Circuit,  wire::Integer, [init_vertex])

Return an iterator over elements on wire.

Start on init_vertex, if supplied, rather than the circuit input vertex.

Qurt.NodeStructs.wireparamelements — Function

wireelements(qc::Circuit,  wire::Integer, [init_vertex])

Return an iterator over elements on wire.

Start on init_vertex, if supplied, rather than the circuit input vertex.

Qurt.NodeStructs.wirevertices — Method

wirevertices(qc::Circuit, wire::Integer)

Return an iterator over ordered vertices on wire beginning with the input node.

Interface

Qurt.Interface — Module

module Interface

This module exports functions that have methods defined in more than one module. It may evolve into a module for functions in the highest API. It's also a bit of a dumping ground at the moment for functions that may better belong elsewhere. For convenience some of the heavily used functions are imported and reexported from the top level of this package.

Qurt.Interface.count_ops — Function

count_ops(obj)

Return a count_map (Dictionary) of the ops in an object.

Qurt.Interface.count_wires — Function

count_wires(qcirc)
count_wires(obj)

Return a count map the number of nodes with (nqu, ncl) qubits and classical bits for each value of the Tuple.

Qurt.Interface.draw — Function

draw(qc::Circuit, args...)

Use Python qiskit to draw qc.

Some arguments arg will work as expected.

Qurt.Interface.isinvolution — Function

isinvolution(obj)

Return true if the obj is the inverse of obj.

Qurt.Interface.node — Function

node(qc, vert)
node(nodes, vert)
node(nodes, verts)

Return the node on vertex vert, or collection verts.

Qurt.Interface.num_clbits — Function

num_clbits(obj)
num_clbits(objs, i)

Return the number of clbits associated with obj or with the ith element of the collection objs.

This function should have a method for any object where it makes sense.

Qurt.Interface.num_inwires — Function

num_inwires(obj)
num_inwires(obj, i)

Return the number of incoming wires (quantum and classical) in obj or in the ith element of the collection objs.

Qurt.Interface.num_outwires — Function

num_outwires(obj)
num_outwires(obj, i)

Return the number of outgoing wires (quantum and classical) in obj or in the ith element of the collection objs.

Qurt.Interface.num_parameters — Function

num_parameters(obj)
num_parameters(obj, i)

Return the number of parameters associated with obj or with the ith element of the collection objs.

This function should have a method for any object where it makes sense.

Qurt.Interface.num_qu_cl_bits — Method

num_qu_cl_bits(obj)
num_qu_cl_bits(objs, i)

Return a Tuple{Int, Int} of number of quantum and classical bits associated with obj or the ith element of objs.

This may be more efficient than calling num_qu_bits and num_cl_bits.

Qurt.Interface.num_qubits — Function

num_qubits(obj)
num_qubits(objs, i)

Return the number of qubits associated with obj or with the ith element of the collection objs.

This function should have a method for any object where it makes sense.

Qurt.Interface.num_wires — Function

num_wires(obj)

Return the number of wires (quantum and classical) in obj or in the ith element of the collection objs.

Qurt.Interface.to_qiskit — Function

to_qiskit(qcircuit)

Convert qcircuit to a Qiskit QuantumCircuit.

You must add package PythonCall to your project and load it before using to_qiskit. Only a relatively small portion of QuantumCircuit is supported.

Qurt.Interface.to_qurt_circuit — Function

to_qurt_circuit(qiskit_circuit)

Convert the Python-qiskit qiskit_circuit to a Qurt.Circuits.Circuit.

Nodes

Qurt.NodeStructs — Module

module NodeStructs

This module Manages data associated with vertices on a DAG. This includes node type, for example io, operator, etc. Also information on which wires pass through/terminate on the node/vertex and which neighboring vertices are on the wires.

The types used here are Node for a single node, and StructVector{<:Node} for a "struct of arrays" collection.

We also have a roll-your-own struct of arrays NodeArray in node_array.jl. It is a bit more performant in many cases. But requires more maintenance.

Qurt.NodeStructs.NodeArray — Type

struct NodeArray

Our custom struct-of-arrays collection of Nodes.

Qurt.Interface.count_elements — Method

count_elements(testfunc::F, nodes::ANodeArrays)

Count circuit elements for which testfunc returns true.

Qurt.NodeStructs.inneighborind — Method

inneighborind(nodes::ANodeArrays, node_ind::Integer, wire::Integer)

Return a Tuple{T,T} of the in-neighbor of node node_ind on wire wire and the wire index of wire on that in-neighbor.

Qurt.NodeStructs.multi_qubit_ops — Method

multi_qubit_ops(nodes::ANodeArrays)

Return a view of nodes containing all with two qubit wires.

Qurt.NodeStructs.n_qubit_ops — Method

n_qubit_ops(nodes::ANodeArrays, n::Integer)

Return a view of nodes containing all with two qubit wires.

Qurt.NodeStructs.named_nodes — Method

named_nodes(nodes::ANodeArrays, names...)

Return a view of nodes containing all with name (Element type) in names

Qurt.NodeStructs.new_node_vector — Function

new_node_vector()

Create an object for storing node information. This includes the element, parameters, information on wires and mapping wires to vertices.

Qurt.NodeStructs.outneighborind — Method

outneighborind(nodes::ANodeArrays, node_ind::Integer, wire::Integer)

Return a Tuple{T,T} of the out-neighbor of node node_ind on wire wire and the wire index of wire on that out-neighbor.

Qurt.NodeStructs.set_inwire_vertex! — Method

set_inwire_vertex!(nodes, src_vertex, wireind, dst_vertex)

Set the inneighbor of src_vertex on wireind to dst_vertex.

Change the inwire map of vertex src_vertex so that wireind maps to dst_vertex.

Qurt.NodeStructs.set_outwire_vertex! — Method

set_outwire_vertex!(nodes, src_vertex, wireind, dst_vertex)

Set the outneighbor of src_vertex on wireind to dst_vertex.

Change the outwire map of vertex src_vertex so that wireind maps to dst_vertex.

Qurt.NodeStructs.two_qubit_ops — Method

two_qubit_ops(nodes::ANodeArrays)

Return a view of nodes containing all with two qubit wires.

Qurt.NodeStructs.wireind — Method

wireind(nodes, node_ind, wire::Integer)

Return the index of wire number wire in the list of wires for node node_ind.

Qurt.NodeStructs.wirevertices — Method

wirevertices(nodes, wire, init_vertex)

Return an iterator over the ordered vertices in nodes on wire beginning with init_vertex.

The final output node is omitted.

Builders

Qurt.Builders — Module

module Builders

This module contains circuit and gate builders implemented as macros

Build gates and circuits with macros @build, @gate, and @gates.

Note that @gate actually constructs not only gates, but any circuit element, such as Measure and Barrier. In particular, the following syntax applies to all circuit elements. (Perhaps @gate should be renamed)

A structure representing a gate G applied at wires (i, [j,...]) is constructed with the syntax G(i, [j,...]). Quantum and classical wires are separated with a semicolon. Gate parameters (p1, [p2,...]) are associated with a gate via curly brackets like this G{p1, [p2,...]}. Both wires and parameters are associated by combining this syntax like this G{p1, [p2,...]}(i, [j,...])

For example

@gate X(1)  # gate and wire
@gate CX(2, 3) # gate and wires
@gate RX{1.5}  # gate and parameters
@gate RX{1.5}(2) # gate, parameters, and wire
@gate Measure(1, 2; 3, 4) # Circuit element and quantum and classical wires

Qurt.Builders.@build — Macro

@build qcircuit gate1 gate2 ...

Add circuit elements to qcircuit.

Qurt.Builders.@gate — Macro

@gate GateName::Element
@gate GateName{param1, [pararam2,...]}
@gate GateName(wire1, [wire2,...])
@gate GateName{param1, [pararam2,...]}(wire1, [wire2,...])

"Build" a gate.

There is no single object that represents a gate application. But it's convenient at times to work with a gate together with its parameters, or the wires that it is applied to. This macro actually packages this information about applying a gate into a struct, which can later be unpacked and inserted into a circuit. For example add_node! accepts types returned by @gate. See also @gates

Qurt.Builders.@gates — Macro

@gates gate1 gate2 ...

Return a Tuple of gates where gates1, gates2, etc. follow the syntax required by @gate. See @gate

Passes

Qurt.Passes — Module

module Passes

At present there are no passes and no pass manager. This module collects circuit transformation and analysis functions that would be used in compiler passes.

Qurt.Passes.cx_cancellation! — Method

cx_cancellation!(qc::Circuit)

Remove CX gates according the the rule CX CX → I.

Replace each sequence of CX gates by one CX gate if the length of the sequence is even, and by nothing if it is odd.

Qurt.Passes.simplify_involution! — Method

simplify_involution!(qc::Circuit, op::Element)

Replace runs of op of odd length by a single op, and remove those of even length.

CompoundGates

Qurt.CompoundGates — Module

module CompoundGates

This module exports the struct CompoundGate.

Qurt.CompoundGates.CompoundGate — Type

CompoundGate{CircuitT}

A gate represented by a unitary circuit.

Wires

Qurt.WiresMod — Module

module WiresMod

This module defines the struct Wires, which tracks data needed to identify wires and their input and output vertices.

Angle

Qurt.Angle — Module

module Angle

This module implements some functions for angles measured in turns. One turn is equal to 2π radians.

Qurt.Angle.Turn — Type

Turn{T}

Angular measurement such that one turn around the circle is equal to one. To convert from radians to Turn divide by 2π.

Qurt.Angle.cos_turn — Method

cos_turn(x)

Compute the cosine of x, an angle measured in "turns".

Qurt.Angle.equal_turn — Method

equal_turn(x, y)

Return true if x and y are equal mod 1.

x and y are intended to be angles measured in "turns".

Qurt.Angle.isapprox_turn — Method

isapprox_turn(x, y; kw...)

Return true if x and y are equal mod 1.

Qurt.Angle.normalize_turn — Method

normalize_turn(x)

Return x mod 1.

The return value is in [0,1). Fast exit if x is already in this range.

Qurt.Angle.sin_turn — Method

sin_turn(x)

Compute the sine of x, an angle measured in "turns".

Qurt.Angle.sincos_turn — Method

sincos_turn(x)

Return a Tuple of the sine and cosine of x, an angle measured in "turns".

Qurt.Angle.tan_turn — Method

tan_turn(x)

Compute the tangent of x, an angle measured in "turns".

IOQDags

Qurt.IOQDAGs — Module

module IOQDAGs

This module exports one function, print_edges.

An extension module that depends on PythonCall allows using Python qiskit to draw circuits with draw.

Qurt.IOQDAGs.print_edges — Method

print_edges([io::IO], qc::Circuit)

Print some information on the edges of qc.

Index

Qurt.Angle

Qurt.Builders

Qurt.Circuits

Qurt.CompoundGates

Qurt.Elements

Qurt.IOQDAGs

Qurt.Interface

Qurt.NodeStructs

Qurt.Passes

Qurt.Qurt

Qurt.WiresMod

Qurt.Elements.Q1GateBlocks

Qurt.Elements.X

Qurt.Elements.Y

Qurt.Angle.Turn

Qurt.Circuits.Circuit

Qurt.Circuits.Circuit

Qurt.Circuits.Circuit

Qurt.CompoundGates.CompoundGate

Qurt.Elements.Element

Qurt.Elements.OpBlock

Qurt.Elements.ParamElement

Qurt.Elements.WiresElement

Qurt.Elements.WiresParamElement

Qurt.NodeStructs.NodeArray

Qurt.WiresMod.Wires

Base.empty

Graphs.inneighbors

Graphs.inneighbors

Graphs.outneighbors

Graphs.outneighbors

Qurt.Angle.cos_turn

Qurt.Angle.equal_turn

Qurt.Angle.isapprox_turn

Qurt.Angle.isapprox_turn

Qurt.Angle.normalize_turn

Qurt.Angle.sin_turn

Qurt.Angle.sincos_turn

Qurt.Angle.tan_turn

Qurt.Circuits.__add_io_node_data!

Qurt.Circuits.add_node!

Qurt.Circuits.check_param_table

Qurt.Circuits.compose

Qurt.Circuits.compose!

Qurt.Circuits.count_ops_longest_path

Qurt.Circuits.global_phase

Qurt.Circuits.insert_node!

Qurt.Circuits.longest_path

Qurt.Circuits.nodes

Qurt.Circuits.num_tensor_factors

Qurt.Circuits.param_map

Qurt.Circuits.param_table

Qurt.Circuits.predecessors

Qurt.Circuits.quantum_successors

Qurt.Circuits.remove_block!

Qurt.Circuits.remove_blocks!

Qurt.Circuits.remove_node!

Qurt.Circuits.successors

Qurt.Circuits.topological_nodes

Qurt.Circuits.topological_vertices

Qurt.Elements._check_begin_block

Qurt.Elements._get_qsyms

Qurt.Elements.inblocks

Qurt.GraphUtils._add_vertex!

Qurt.GraphUtils._add_vertices!

Qurt.IOQDAGs.print_edges

Qurt.Interface.check

Qurt.Interface.count_elements

Qurt.Interface.count_op_elements

Qurt.Interface.count_ops

Qurt.Interface.count_ops_vertices

Qurt.Interface.count_wires

Qurt.Interface.depth

Qurt.Interface.draw

Qurt.Interface.getparam

Qurt.Interface.iscustomgate

Qurt.Interface.isinvolution

Qurt.Interface.node

Qurt.Interface.num_clbits

Qurt.Interface.num_clbits

Qurt.Interface.num_inwires

Qurt.Interface.num_outwires

Qurt.Interface.num_parameters

Qurt.Interface.num_parameters

Qurt.Interface.num_qu_cl_bits

Qurt.Interface.num_qubits

Qurt.Interface.num_qubits

Qurt.Interface.num_wires

Qurt.Interface.to_qiskit

Qurt.Interface.to_qurt_circuit

Qurt.NodeStructs.find_nodes

Qurt.NodeStructs.inneighborind

Qurt.NodeStructs.multi_qubit_ops

Qurt.NodeStructs.n_qubit_ops

Qurt.NodeStructs.named_nodes

Qurt.NodeStructs.new_node_vector

Qurt.NodeStructs.one_qubit_ops

Qurt.NodeStructs.outneighborind

Qurt.NodeStructs.rewire_across_node!

Qurt.NodeStructs.set_inwire_vertex!

Qurt.NodeStructs.set_outwire_vertex!

Qurt.NodeStructs.two_qubit_ops

Qurt.NodeStructs.unpackwires

Qurt.NodeStructs.wireelements

Qurt.NodeStructs.wireind

Qurt.NodeStructs.wireind

Qurt.NodeStructs.wireparamelements

Qurt.NodeStructs.wirevertices

Qurt.NodeStructs.wirevertices

Qurt.Passes.cx_cancellation!

Qurt.Passes.simplify_involution!

Qurt.WiresMod.input_vertex

Qurt.WiresMod.output_vertex

Qurt.Builders.@build

Qurt.Builders.@gate

Qurt.Builders.@gates

Qurt.Elements.@new_elements