Internals

Circuit Elements

Qurt.Angle.isapprox_turn — Method

isapprox_turn(x::ParamElement, y::ParamElement; kw...)

Return true if x and y are approximately equal. The element types must be equal. Angle.isapprox_turn must return true element-wise on the parameters.

Qurt.Elements.@new_elements — Macro

@new_elements BlockName sym1 sym2 ...

Add new circuit element symbols to the block of elements named BlockName.

Examples

julia> @new_elements MiscGates MyGate1 MyGate2

Circuits

Base.empty — Method

empty(qc::Circuit)

Return an object that is a copy of qc except that all circuit elements other than input and output nodes are not present.

Qurt.Circuits.__add_io_node_data! — Method

__add_io_node_data!(graph, nodes, nqubits, nclbits)

Add input and output nodes to nodes. Wires numbered 1 through nqubits are quantum wires. Wires numbered nqubits + 1 through nqubits + nclbits are classical wires.

Qurt.Circuits.check_param_table — Method

check_param_table(qc::Circuit)

Check that the (node, paramter position) pairs recorded in the table actually contain the recorded parameter. This check could fail if reindexing is not done properly.

Qurt.Interface.check — Method

check(qc::Circuit)

Throw an Exception if any of a few checks on the integrity of qc fail.

Qurt.Interface.count_op_elements — Method

count_op_elements(qc::Circuit)

Return the number of circuit elements that are not IO nodes. This should be the number that are operation or instruction nodes.

Examples

julia> import Qurt

julia> using Qurt.Circuits: Circuit, count_op_elements

julia> using Qurt.Builders: @build

julia> using Qurt.Elements: H, CX

julia> qc = Circuit(2);

julia> count_op_elements(qc)
0

julia> @build qc H(1) CX(1, 2);

julia> count_op_elements(qc)
2

Qurt.Interface.depth — Method

depth(qc::Circuit)

Compute the depth of qc.

Qurt.Interface.num_clbits — Method

num_clbits(qc::Circuit)

Return the number of classical bits in qc.

Qurt.Interface.num_parameters — Method

num_parameters(qc::Circuit)

Return the number of unique symbolic parameter expressions in use in gates in qc.

For example, if exactly t1, t2, and t1 - t2 are present, then num_parameters will return 3. Non-symbolic parameters, such as 1.5, do not count.

Qurt.Interface.num_qubits — Method

num_qubits(qc::Circuit)

Return the number of qubits in qc.

Qurt.NodeStructs.wireind — Method

wireind(circuit, node_ind, wire::Integer)

Return the index of wire number wire in the list of wires for node node_ind.

Interface

Qurt.Interface.iscustomgate — Function

iscustomgate(gate)

Return true if the gate is tagged CustomGate::Element.

Nodes

Graphs.inneighbors — Method

inneighbors(circuit, node_ind::Integer, wire::Integer)

Return the node index connected to node_ind by incoming wire number wire.

Graphs.inneighbors — Method

inneighbors(circuit, node_ind::Integer)

Return collection of incoming neighbor nodes in wire order.

Nodes may appear more than once if they are connected by multiple wires.

Graphs.outneighbors — Method

outneighbors(circuit, node_ind::Integer, wire::Integer)

Return the node index connected to node_ind by outgoing wire number wire.

Graphs.outneighbors — Method

outneighbors(circuit, node_ind::Integer)

Return collection of outgoing neighbor nodes in wire order.

Nodes may appear more than once if they are connected by multiple wires.

Qurt.Interface.count_ops_vertices — Method

count_ops_vertices(nodes::ANodeArrays, vertices)

Return a count map of elements on vertices.

Qurt.Interface.getparam — Method

getparam(nodes::ANodeArrays, ind::Integer, pos::Integer)

The the posth parameter at node index ind in nodes.

Qurt.NodeStructs.find_nodes — Method

find_nodes(testfunc::F, nodes::NodeVector, fieldname::Symbol) where {F}
find_nodes(testfunc::F, nodes::NodeVector, fieldnames::Tuple) where {F}
find_nodes(testfunc::F, nodes::NodeVector, ::Val{fieldnames}) where {F, fieldnames}

Return a view of nodes filtered by testfunc.

testfunc must take a single argument. It will be passed an structure with fields fieldnames. Only nodes for which testfunc returns true will be kept.

Calling find_nodes with fewer fields is more performant.

Examples

Find two qubit operations.

julia> find_nodes(x -> x.numquwires == 2, nodes, :numquwires)

Qurt.NodeStructs.one_qubit_ops — Method

one_qubit_ops(nodes::ANodeArrays)

Return a view of nodes containing all with two qubit wires.

Qurt.NodeStructs.rewire_across_node! — Method

rewire_across_node!(nodes::ANodeArrays, vind::Integer)

Wire incoming neighbors of vind to outgoing neighbors of vind, preserving the order of wires on ports.

Qurt.NodeStructs.unpackwires — Method

unpackwires(wires, nqu::Integer)

Return a tuple of two tuples, where the first contains the quantum wires and the second contains the classical wires.

wires contains all of the wires, with quantum wires first.

Builders